requirements.txt
colorlog==6.7.0
Flask==2.3.2
Flask_Cors==4.0.0
flask_sqlalchemy==3.0.5
openai==0.27.8
python-dotenv==1.0.0
Requests==2.31.0
semantic_kernel==0.3.3.dev0
SQLAlchemy==2.0.17
Werkzeug==2.3.6
gunicorn==20.1.0
pymssql==2.2.7
Flask-Session==0.5.0
backoff==2.2.1
pre-commit>=3.4